Merge bk://are.twiddle.net/axp-2.5/ into home.transmeta.com:/home/torvalds/v2.5/linux
2003/06/19 08:49:10-07:00 torvalds
Merge bk://linux-pnp.bkbits.net/pnp-2.5
into home.transmeta.com:/home/torvalds/v2.5/linux

2003/06/18 22:38:55+00:00 ambx1
[PNP] Important Resource Parsing Fixes

In some cases, we're reading the wrong bits for large tags.  This patch corrects
the issue by setting the affected bits forward by an offset of 2 (skipping over
the size portion of the tag).

2003/06/18 22:37:29+00:00 ambx1
[PNP] Remove some leftover resource config options in isapnp

Must have missed it earlier, but the pci module parameter is not needed.

2003/06/18 22:36:19+00:00 ambx1
[PNP] Trivial Typo fix regarding DMAs

The irq index is used instead of the dma index when parsing dmas.

2003/06/18 22:34:14+00:00 ambx1
[PNP] re-add the previously removed "get" command in interface.c.

This patch adds the "get" command because at this point it is needed
for debugging.

2003/06/18 22:31:55+00:00 ambx1
[PNP] PnPBIOS resource setting fix

If a device is disabled when initially read, its blank resource data will not
be cleared and the pnp layer will assume incorrectly that the device has
already been configured.  This patch resolves the issue by initializing the
resource table if the device is found to be disabled.

2003/06/18 22:28:13+00:00 ambx1
[PNP] Module Compilation Fix

Fixes a trivial typo in an export symbol macro.

2003/06/18 22:26:10+00:00 ambx1
[PNP] /drivers/pnp/resource.c check_region warning fix

This patch resolves the compiler warning caused by the depreciated check_region
function.  It may not be the best solution but check_region really is what is
needed here because we never actually have to call "request_region".  If prefered,
I could alternatively request and release but doing so would be less efficient.

2003/06/18 22:23:14+00:00 ambx1
[PNP] Resource Management Cleanups and Updates

This patch does the following...
1.) changes struct pnp_resources to pnp_option for clarity
2.) greatly cleans up resource option registration
3.) removes some of the current conflict prevention code in
order to increase flexibility, (users will have more control)
4.) various manager cleanups, resulting code is more efficient
5.) fixes the locking bugs many have reported (now uses a mutex)
6.) removes the conflict displaying interface
 - it is better to handle such things in user space
7.) also many misc. cleanups

BKrev: 3ef485d4dA3xJHfdWeaz03h2GwigUw