[PATCH] JBD: honour read-only mounts more carefully
From: "Stephen C. Tweedie" <sct@redhat.com>

ext3 has long had a problem wherein it will unnecessarily write to a
read-only filesystem during the mount process.  It does this in preparing the
journal superblock's sequence numbers.

But if the filesystem was shut down cleanly we do not need to do this.
Detect the situation and avoid modifying and writing out the journal
superblock.

BKrev: 3ef48610WUjU4Fwvjgbj8OGhCkVCUg