Merge davem@nuts.ninka.net:/home/davem/src/BK/sparc-2.5 into kernel.bkbits.net:/home/davem/sparc-2.5
2003/06/22 18:40:55-07:00 ldl
[PATCH] fix nbd driver for 2.5 block layer

This makes NBD work with the new linux 2.5 block layer design.
Specifically, it fixes memory corruption that results from module
removal and possible memory corruption from sending or receiving disk
data from the server.

It essentially rolls together the changes from two of the last patchlets
that I emailed: the fix for module removal & the fix for incorrect
struct bio usage.  I believe it's wisest to roll these both together
into this one patch since they both deal with making NBD work better
with the 2.5 linux block layer design and without either of which, it's
possible that NBD will corrupt memory.

Other changes I'd like to see introduced (like in the earlier jumbo
patch) meanwhile are feature enhancements so they can wait.  This patch
also should address all the very helpful concerns that have been raised
so far.  Particularly:

1. that the very first submitted NBD patch was broken down [Andrew]
2. that only 1 spinlock is used for all the NBD request_queue structures
   used [Jens,Al]
3. that kmap() is used in case of highmem pages [Jens]
4. that the allocation of request_queue is dynamic and seperate from
   other allocated objects [Al]

2003/06/22 18:37:32-07:00 Andries.Brouwer
[PATCH] loop.c cleanups

This does the following:
 - remove trailing spaces
 - make loop.h independent by including bio.h, blk.h, spinlock.h
 - replace the lock/unlock functions by module_get/module_put;
   in struct loop this is the change
	-       void (*lock)(struct loop_device *);
	-       void (*unlock)(struct loop_device *);
	+       struct module *owner;
 - replace the integer lo_encrypt_type by the pointer lo_encryption;
   there was a race with loop_unregister_transfer
 - fixed an off-by-one in loop_register_transfer

This is Step 1 of a series of half a dozen or so.

Half of the above is from Jari. Anything that is wrong is mine.

BKrev: 3ef66cbdwUOstAOYXosXis88XvyBxg