[PATCH] Include update mode in declaration of RPC information caches.
The sunrpc/cache.c caches which store nfsd export information
can be updated either "inplace" or by replacing the entry.

replacement is needed when an entry hold a reference to some
other object, so the reference counts work properly.

"inplace" can be used when no such references are held, and should be
used when the object could be refered to by another cache (as otherwise
the other cache would have to be updated whenever this one is).

Previously the type of update (inplace or replace) was specified
as an argument to the *_lookup operation.  This too easily lead to
inconsistancies.

With this patch, the update mode is specified when the cache is
declared.

BKrev: 3ef8a321Pv0UXjMaGLh0SavdclhxAg