[PATCH] Add some tracing when showing the content of an RPC cache.
The /proc/net/rpc/*/content files now will show
entries that are still in the cache, but are either
expired or negative, as comment.

ip_map_show is enhance to work if called with a negative
or incomplete entry.

Also if cache debugging is enabled, the expiry time and
refcount of each entry will be included in a comment.

BKrev: 3ef8a32bIYH1uOy-MyM-dHHhIVnGSw