Merge bk://kernel.bkbits.net/jgarzik/net-drivers-2.5 into home.osdl.org:/home/torvalds/v2.5/linux
2003/06/25 11:52:46-07:00 rddunlap
[PATCH] Unionize IO-APIC registers

[ Registers of the world, unite! ]

This makes the IO-APIC data structures use unions, so that we can
cleanly access them both as flat "raw" values, and as the bitmap
sub-entries.

2003/06/25 08:59:39-07:00 torvalds
Merge bk://ppc.bkbits.net/for-linus-ppc
into home.transmeta.com:/home/torvalds/v2.5/linux

2003/06/25 22:52:27+10:00 paulus
PPC32: remove check for ERESTARTNOHAND in syscall return path.

This error is handled in the signal delivery code and should never be
returned from a syscall unless a signal is pending.  Grepping seems to
indicate that that is in fact the case (but not for ERESTARTSYS, but
that is another problem).

2003/06/24 21:51:11-07:00 davem
[SPARC64]: Update defconfig.

2003/06/24 21:50:45-07:00 davem
[SERIAL]: Sanitize sparc serial console configuration.

2003/06/24 21:12:20-07:00 davem
[SPARC64]: Update defconfig.

2003/06/24 21:10:34-07:00 davem
[PCI]: Export pci_enable_device_bars to modules.

2003/06/24 21:08:49-07:00 davem
[USB]: Use linux/dma-mapping.h not asm/dma-mapping.h in kaweth.c

BKrev: 3efa1cb2jZJtC_AeZ-aFjo3U26PuZA