[PATCH] Use Local Percpu Macros for Local Percpu Variables
In general, it is more better to use get_cpu_var() and __get_cpu_var()
to access per-cpu variables on this CPU than to use smp_processor_id()
and per_cpu().  In the current default implemention they are equivalent,
but on IA64 the former is already faster, and other archs will follow.

BKrev: 3efc744bipCUklN9SP35mBSslkhvnw