[PATCH] Remove racy check_mem_region() call from arc-rimi.c
From: Bob Miller <rem@osdl.org>

Removed the check_mem_region() call and replaced with request_mem_region().

Because of the way the driver is structured the first request_mem_region()
call gets the default memory area.  After probing the complete memory area
that is needed to communicate with the device is known, so the first memory
area is released and the complete area is requested.

BKrev: 3efe0447qUTDNqgFxP1rIqO4q9EPig