Merge bk://kernel.bkbits.net/gregkh/linux/linus-2.5 into home.osdl.org:/home/torvalds/v2.5/linux
2003/07/01 14:13:35-07:00 greg
Merge kroah.com:/home/greg/linux/BK/bleed-2.5
into kroah.com:/home/greg/linux/BK/gregkh-2.5

2003/07/01 13:45:37-07:00 stern
[PATCH] USB: Reconcile unusual_devs.h in 2.4 and 2.5

This patch for 2.5 copies into unusual_devs.h an entry that is present in
2.4 but missing here.  It also adds back the IGNORE_SER flag with a dummy
value, just for backward compatibility.

2003/07/01 13:45:23-07:00 lkml001
[PATCH] USB: usb-skeleton.c usb_buffer_free() not called

dev->udev was set to NULL to indicate a device disconnect but we need
this value for usb_buffer_free() when device is still opened and cleanup is
delayed until skel_release().
I've added a dev->present for preventing device read, write and ioctl.
Also URB_NO_DMA_MAP is changed to URB_NO_TRANSFER_DMA_MAP and a unused
var is removed.

2003/07/01 13:30:40-07:00 greg
[PATCH] USB: move ohci's sysfs files to the class device instead of the pci device.

2003/07/01 13:30:27-07:00 greg
[PATCH] USB: move ehci's sysfs files to the class device instead of the pci device.

2003/07/01 13:30:15-07:00 greg
[PATCH] USB: make struct usb_bus a struct class_device

This creates sys/class/usb_host/ which lists all usb hosts in the system.

2003/06/30 17:16:53-07:00 greg
[PATCH] USB: turn down some debugging messages in uhci-hcd.

2003/06/30 11:25:20-07:00 greg
Merge kroah.com:/home/greg/linux/BK/bleed-2.5
into kroah.com:/home/greg/linux/BK/gregkh-2.5

2003/06/30 11:18:01-07:00 grigouze
[PATCH] USB: zaurus SL-C700

This is a patch for usbnet for working with Zaurus SL-C700.
The productid is different from other Zaurus, so i add an entry for it :)

2003/06/30 11:01:44-07:00 mdharm-usb
[PATCH] USB storage: logic cleanup

This patch cleans up the logic involved in setting URB flags a bit so it's
more evident that some flags are always set and other flags are only set
under certain conditons.

2003/06/30 11:00:45-07:00 mdharm-usb
[PATCH] USB storage: General purpose I/O buffer allocation and management

This patch makes our private I/O buffer allocated such that it's pre-mapped
for DMA.  We then add some logic to make sure that we don't try to re-map
it.

We also make the size of the buffer large enough for other sub-drivers,
which will be converted shortly.

2003/06/30 10:59:49-07:00 mdharm-usb
[PATCH] USB storage: Cosmetic cleanups

This is some cosmetic cleanups.  Things like changing dr to cr (they used
to stand for usb_devrequest, which is now usb_ctrlrequest), using strlcpy
instead of strncpy, and modifying some comments.

2003/06/30 10:59:06-07:00 mdharm-usb
[PATCH] USB storage: create private I/O buffer

This patch creates a private I/O buffer for the 'core' subdrivers.  This
allows us to fix all cases of DMAing to/from the stacks, statics, or other
bad places.

2003/06/30 10:58:29-07:00 mdharm-usb
[PATCH] USB storage: avoid sending URBs when disconnecting

This patch avoids sending URBs which are part of the reset system when the
device is being disconnected.  It also makes the control thread avoid
starting anything for a device in that state.

2003/06/30 10:58:01-07:00 mdharm-usb
[PATCH] USB storage: create associate_dev(), more US_*_DEVICE printout

This patch creates an associate_dev(), which is the mirror to
disassociate_dev().  It also adds some more printout when identifiying
devices that qualify for US_*_DEVICE codes so we need less interaction with
the reporting user.

2003/06/30 10:57:07-07:00 mdharm-usb
[PATCH] USB storage: unusual_devs.h cleanups

This patch re-sorts unusual_devs.h and removes some duplicate entries.

2003/06/27 11:55:31-07:00 greg
Cset exclude: cweidema@indiana.edu|ChangeSet|20030621162021|08529

BKrev: 3f021176EkEmX9fL6BH3JQ2z5H12PQ