[PATCH] Set limits on CONFIG_LOG_BUF_SHIFT
From: bert hubert <ahu@ds9a.nl>

Attached patch adds a range check to LOG_BUF_SHIFT and clarifies the
configuration somewhat.  I managed to build a non-booting kernel because I
thought 64 was a nice power of two, which lead to the kernel blocking when
it tried to actually use or allocate a 2^64 buffer.

BKrev: 3f03c413aIM-TViUrBYPxUcIzEJmSw