Merge bk://kernel.bkbits.net/jgarzik/net-drivers-2.5 into home.osdl.org:/home/torvalds/v2.5/linux
2003/07/10 19:32:37-07:00 torvalds
Merge bk://kernel.bkbits.net/jgarzik/irda-2.5
into home.osdl.org:/home/torvalds/v2.5/linux

2003/07/10 14:59:35-07:00 greg
Merge kroah.com:/home/greg/linux/BK/bleed-2.5
into kroah.com:/home/greg/linux/BK/gregkh-2.5

2003/07/10 17:49:22-04:00 jt
[IrDA] nsc 39x support

		<Original patch from Jan Frey>
	o [FEATURE] Add preliminary support for NSC PC8739x chipset
		(IBM R31 laptops)

2003/07/10 17:48:50-04:00 jt
[IrDA] irda-usb endian

	o [FEATURE] Add USB-ID of new device
		<Original patch from Jacek Jakubowski>
	o [CORRECT] Endianess fixes (for PPC and co.)

2003/07/10 17:48:17-04:00 jt
[IrDA] setup dma fix

	o [CORRECT] Make ISA drivers depend on ISA. This is the consequence
		from David's change to setup_dma().
	o [CORRECT] Make new dongle drivers depend on sir-dev (they require it)
	o [FEATURE] Make old dongle drivers depend on irtty/irport
	o [FEATURE] irda-usb driver is no longer experimental

2003/07/10 17:47:46-04:00 jt
[IrDA] IrCOMM devfs

		<Original patch from Martin Diehl>
	o [CRITICA] fix IrCOMM bogus device names with devfs

2003/07/10 17:47:15-04:00 jt
[IrDA] irnet cast

		<Suggested by David S. Miller>
	o [FEATURE] remove pointer casting in IrNET debug code missed by David.

2003/07/10 17:46:44-04:00 jt
[IrDA] irtty leaks

		<Stanford checker>
	o [CORRECT] fix two additional potential skb leaks in IrTTP.

2003/07/10 17:46:11-04:00 jt
[IrDA] struct check

		<Suggested by Russell King>
	o [FEATURE] Add struct size check for buggy compilers

2003/07/10 17:45:39-04:00 jt
[IrDA] include cleanup

	o [CORRECT] Cleanup cruft from <linux/irda.h>

2003/07/10 13:18:12-07:00 greg
[PATCH] USB: remove pointless warning about using usbdevfs

This is being printed when it shouldn't, but there really isn't an
easy way to fix it.  This closes bug #223.

2003/07/09 21:51:31-07:00 stern
[PATCH] USB: Updates for unusual_devs.h

These changes to unusual_devs.h can be applied to both 2.5 and 2.4.  They
reflect reports from users in response to the new kernel log messages
about unneeded entries in the file.

2003/07/09 21:05:29-07:00 greg
[PATCH] USB: fix up my USB Bluetooth entry to help prevent confusion in the future.

2003/07/09 20:52:35-07:00 david-b
[PATCH] USB: usbnet, don't NET_XMIT_DROP

Martin Diehl pointed out that the network layer treats
a tx drop notification in special ways, so "usbnet"
shouldn't be using them.  This fixes that problem,
which could potentially explain some infrequent oopsing
seen in the network layer.  (With memory so tight that
an URB can't be gotten.)

Please merge.

2003/07/09 20:52:04-07:00 david-b
[PATCH] USB: usb_get_string(), don't use bogus ids

This fixes a problem that's been around for some time:
usbcore will sometimes use un-initialized memory as if
it holds a valid language code.

2003/07/09 20:51:38-07:00 stern
[PATCH] USB: Small correction to usb-skeleton.c

This fixes a minor error in usb-skeleton's disconnect() routine: if the
interface's private data is NULL, the current code exits without releasing
the disconnect_sem semaphore.

I removed the test entirely because I can't think of any situation where
that private data actually would be NULL, other than a pretty badly
malfunctioning system.  Why test for something that should never happen?
And if it does happen, we shouldn't want the disconnect routine to fail
silently -- we should want to see a nice big segfault (when the NULL
pointer is dereferenced) so that we can find and fix the underlying error.

Is there any sort of convention (a la Documentation/CodingStyle) about
whether this approach should be used in general?  There are _lots_ of
places in the kernel where unnecessary checks for NULL pointers are made.

2003/07/09 20:51:05-07:00 mdharm-usb
[PATCH] USB: fix datafab and freecom to use I/O buffer

This patch makes the Datafab and the Freecom driver both use the I/O safe
buffer us->iobuf instead of local stack or local allocations.

2003/07/09 20:50:32-07:00 mdharm-usb
[PATCH] USB: fix usb-storage initializers

This patch makes the 2.5 initializers use the DMA-safe us->iobuf instead of
local stack or local dynamic allocation.

2003/07/09 20:49:56-07:00 abbotti
[PATCH] USB: ftdi_sio update

A couple of patches for the ftdi_sio driver on 2.5.73

"2.5.73-ftdi_sio-bigupdate.patch" - this is a big, ugly patch to
bring the ftdi_sio driver in line with what recently submitted for
inclusion in the 2.4.x kernel
(v2.4/testing/cset/cset-abbotti@mev.co.uk|ChangeSet|20030627190202|35624.txt),
but with necessary changes for 2.5.x.  I've bumped the driver's
version number to 1.4.0 to distinguish it from that submitted for
the 2.4.x kernel.

"2.5.73-ftdi_sio-dynamicwrite.patch" - this adds to the above patch
to replace the pool of urbs used for writes with dynamically
allocated urbs and transfer buffers.  I've tested this on the
FT8U232AM and FT232BM and I think I've catered for the older "SIO"
device, but I don't have a device to test that.  Also, it is not
safe to unload the driver module while write urbs are in flight
(same as for the Visor driver).

2003/07/09 17:39:41-07:00 lkml001
[PATCH] usblp: usb_buffer_free() not called
Here is the blind flight :-)
===
drivers/usb/class/usblp.c
usblp->dev was set to NULL to indicate a device disconnect but we need
this value for usb_buffer_free() when device is still opened and cleanup
is delayed until usblp_release().
We have a usblp->present now for preventing device read, write, open and ioctl.

BKrev: 3f0e21e4YokVMPkANm9h1PWZJtreXA