[PATCH] Ext3 xattr credits fix for quotas
From: Andreas Gruenbacher <agruen@suse.de>

The xattr and acl code are not properly reserving credits for quotas.
EXT3_DATA_TRANS_BLOCKS is an overestimate of the credits required including
quotas.  Make it a little more tight, and use it in the xattr and acl code
to be quota safe.

BKrev: 3f181f7b-M7nPYTu0kcMB2U95YJ2qQ