[PATCH] fix return of compat_sys_sched_getaffinity
From: Anton Blanchard <anton@samba.org>

On success getaffinity should return the size of mask.  It looks like
Randy's patch (which fixed a bogus return) missed this bit.

In reality I'm not sure if we care, and with cpu bitmasks this stuff is
going to break.

BKrev: 3f181fb0by2EewTGcToC4QbOWN0jDA